% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ace_snackfoods}
\alias{ace_snackfoods}
\title{Ace Snackfoods Kiwi Bubbles trial}
\format{A named list of two data frames:
\describe{
\item{\code{calibration}}{Table 1: Cumulative number of households that have made a trial purchase by the end of weeks 1–24.}
\item{\code{validation}}{Table A1: Cumulative number of households that have made a trial purchase by the end of weeks 25–52.}
}}
\usage{
data(ace_snackfoods)
}
\description{
Ace Snackfoods, Inc. has developed a new shelf-stable juice product called
Kiwi Bubbles. Sales of the new product are being monitored in a limited
geographic area over a year through a market research panel that follows
the purchases of a representative sample of 1499 households at supermarkets,
convenience stores, warehouse clubs, and mass merchandisers in the test
area. The product has now been under test for 24 weeks.
}
\details{
The marketing manager for Kiwi Bubbles would like a forecast of the product’s
year-end performance in the test market. As a first step, she wants a forecast
of the number of households that will have made a trial purchase by Week 52.
}
\references{
Peter S. Fader, Bruce G. S. Hardie, Daniel McCarthy & Ramnath
Vaidyanathan (2019) Exploring the Equivalence of Two Common Mixture Models
for Duration Data, The American Statistician, 73:3, 288-295, DOI:
10.1080/00031305.2018.1543134
\href{https://doi.org/10.1080/00031305.2019.1543134}{Web}
}
\keyword{datasets}
