\name{BTYD2-package}
\alias{BTYD2-package}
\alias{BTYD2}
\docType{package}
\title{
\packageTitle{BTYD2}
}
\description{
\packageDescription{BTYD2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BTYD2}
\packageIndices{BTYD2}
}

\author{
\packageAuthor{BTYD2}

Maintainer: \packageMaintainer{BTYD2}
}

\references{

See \url{www.brucehardie.com} for papers, notes, and datasets relating
to applied probability models in marketing.

Fader, Peter S., and Bruce G.S. Hardie. \dQuote{A Note on Deriving the
Pareto/NBD Model and Related Expressions.}
November. 2005. Web. \url{http://www.brucehardie.com/notes/008/}

Fader, Peter S., Bruce G.S. Hardie, and Ka L. Lee. \dQuote{RFM and CLV:
  Using Iso-Value Curves for Customer Base Analysis.} \emph{Journal of
  Marketing Research} Vol.42, pp.415-430. November. 2005. \url{http://www.brucehardie.com/papers.html}

Fader, Peter S., and Bruce G.S. Hardie. \dQuote{Deriving an Expression
  for P (X(t) = x) Under the Pareto/NBD Model.}
September. 2006. Web. \url{http://www.brucehardie.com/notes/012/}

Fader, Peter S., and Bruce G.S. Hardie. \dQuote{Creating an RFM summary
  using Excel.} December. 2008. Web. \url{http://www.brucehardie.com/notes/022/}

Fader, Peter S., Bruce G.S. Hardie, and Jen Shang. \dQuote{Customer-Base
  Analysis in a Discrete-Time Noncontractual Setting.} \emph{Marketing
  Science} 29(6),
  pp. 1086-1108. 2010. INFORMS. \url{http://www.brucehardie.com/papers/020/}

  Jerath, Kinshuk, Peter S. Fader, and Bruce
G.S. Hardie. \dQuote{Customer-Base Analysis on a 'Data Diet': Model
Inference Using Repeated Cross-Sectional Summary (RCSS) Data.}
June. 2011. Available at SSRN: \url{http://ssrn.com/abstract=1708562} or
\url{http://dx.doi.org/10.2139/ssrn.1708562}

Fader, Peter S., Bruce G.S. Hardie, and Ka L. Lee. \dQuote{``Counting Your Customers'' the Easy Way:
An Alternative to the Pareto/NBD Model.} \emph{Marketing Science} Vol.24, pp.275-284. Spring. 2005. 
\url{http://www.brucehardie.com/papers.html}

Fader, Peter S., Hardie, Bruce G.S., and Lee, Ka Lok. \dQuote{Computing P(alive) Using
the BG/NBD Model.} December. 2008. Web. \url{http://www.brucehardie.com/notes/021/palive_for_BGNBD.pdf}

}

\keyword{ package }

