% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi.R
\name{equi.bg.LL}
\alias{equi.bg.LL}
\title{Beta-geometric log-likelihood function}
\usage{
equi.bg.LL(params, x, t, n = 1499)
}
\arguments{
\item{params}{A vector of the two Beta-Geometric parameters gamma and delta.}

\item{x}{A vector with the number of transactions at discrete times t.}

\item{t}{The timeline of transactions in vector x.}

\item{n}{The total possible number of transactions at time t.}
}
\value{
The log likelihood of observing the combination of transactions \code{x}
over timeline \code{t} given that \code{x} follows a Beta-Geometric distribution with
parameters \code{params}.
}
\description{
Log-likelihood function for the BG
}
\details{
The log-likelihood function is a building block for estimating the value of
\code{params} given a buying pattern \code{x} observed over time \code{t}. See the example.
}
\examples{
# Get the calibration data. Remember that
# counts of households are cumulative, so 
# turn them to incremental:
library(tidyverse)
data("ace_snackfoods")
data <- ace_snackfoods$calibration \%>\% 
  rename(x = `Number of households`) \%>\% 
  mutate(bought = x - lag(x)) \%>\% 
  mutate(bought = if_else(is.na(bought), x, bought))
  
# Get maximum likelihood estimates:   
mlset <- optim(par = c(0, 3), 
               fn = equi.bg.LL, 
               x = data$bought, 
               t = data$Week, 
               n = 1499, 
               method="BFGS", 
               control=list(fnscale=-1))

# Log-likelihood maximum:
LL <- mlset$value

# Parameter estimates:
# Recall that bg.LL exponentiates par in
# order to enforce gamma > 0, delta > 0.
params <- exp(mlset$par)
}
\references{
Peter S. Fader, Bruce G. S. Hardie, Daniel McCarthy & Ramnath
Vaidyanathan (2019) Exploring the Equivalence of Two Common Mixture Models
for Duration Data, The American Statistician, 73:3, 288-295, DOI:
10.1080/00031305.2018.1543134
\href{https://doi.org/10.1080/00031305.2019.1543134}{Web}
}
